+++ prp/parse/virulencefinder.py
@@ -18,51 +18,57 @@ LOG = logging.getLogger(__name__)
 
 
 def parse_vir_gene(
-    info: dict[str, Any], subtype: ElementVirulenceSubtype = ElementVirulenceSubtype.VIR
+    info: dict[str, Any], function: str, subtype: ElementVirulenceSubtype = ElementVirulenceSubtype.VIR
 ) -> VirulenceGene:
     """Parse virulence gene prediction results."""
-    start_pos, end_pos = map(int, info["position_in_ref"].split(".."))
-    # Some genes doesnt have accession numbers
-    accnr = None if info["accession"] == "NA" else info["accession"]
+    accnr = info.get("ref_acc", None)
+    if accnr == "NA":
+        accnr = None
     return VirulenceGene(
         # info
-        gene_symbol=info["virulence_gene"],
+        gene_symbol=info["name"],
         accession=accnr,
-        sequence_name=info["protein_function"].strip(),
+        sequence_name=function,
         # gene classification
         element_type=ElementType.VIR,
         element_subtype=subtype,
         # position
-        ref_start_pos=start_pos,
-        ref_end_pos=end_pos,
-        ref_gene_length=info["template_length"],
-        alignment_length=info["HSP_length"],
+        ref_start_pos=int(info["ref_start_pos"]),
+        ref_end_pos=int(info["ref_end_pos"]),
+        ref_gene_length=int(info["ref_seq_length"]),
+        alignment_length=int(info["alignment_length"]),
         # prediction
-        identity=info["identity"],
-        coverage=info["coverage"],
+        identity=float(info["identity"]),
+        coverage=float(info["coverage"]),
     )
 
 
-def _parse_vir_results(pred: str) -> VirulenceElementTypeResult:
+def _parse_vir_results(pred: dict[str, Any]) -> VirulenceElementTypeResult:
     """Parse virulence prediction results from virulencefinder."""
-    # parse virulence finder results
-    species = list(k for k in pred["virulencefinder"]["results"])
     vir_genes = []
-    for key, genes in pred["virulencefinder"]["results"][species[0]].items():
+
+    phenotypes = pred.get("phenotypes", {})
+    seq_regions = pred.get("seq_regions", {})
+    
+    for key, pheno in phenotypes.items():
+        function = pheno.get("function")
+        ref_dbs = pheno.get("ref_database", [])
+
         # skip stx typing result
-        if key == "stx":
+        if any("stx" in db for db in ref_dbs):
             continue
+
         # assign element subtype
-        virulence_group = key.split("_")[1] if "_" in key else key
-        match virulence_group:
-            case "toxin":
-                subtype = ElementVirulenceSubtype.TOXIN
-            case _:
-                subtype = ElementVirulenceSubtype.VIR
+        subtype = ElementVirulenceSubtype.VIR
+        if any("toxin" in db for db in ref_dbs):
+            subtype = ElementVirulenceSubtype.TOXIN
+
         # parse genes
-        if not genes == "No hit found":
-            for gene in genes.values():
-                vir_genes.append(parse_vir_gene(gene, subtype))
+        for region_key in pheno.get("seq_regions", []):
+            seq_info = seq_regions.get(region_key)
+            if not seq_info:
+                continue
+            vir_genes.append(parse_vir_gene(seq_info, subtype=subtype, function=function))
     # sort genes
     genes = sorted(vir_genes, key=lambda entry: (entry.gene_symbol, entry.coverage))
     return VirulenceElementTypeResult(genes=genes, phenotypes={}, variants=[])
@@ -79,7 +85,7 @@ def parse_virulence_pred(path: str) -> VirulenceMethodIndex | None:
     LOG.info("Parsing virulencefinder virulence prediction")
     with open(path, "rb") as inpt:
         pred = json.load(inpt)
-        if "virulencefinder" in pred:
+        if "seq_regions" in pred and "phenotypes" in pred: # Aim: check if file is empty or if it comes from the right tool?
             results: VirulenceElementTypeResult = _parse_vir_results(pred)
             result = VirulenceMethodIndex(
                 type=ElementType.VIR, software=Software.VIRFINDER, result=results
@@ -96,25 +102,26 @@ def parse_stx_typing(path: str) -> MethodIndex | None:
         pred_obj = json.load(inpt)
         # if has valid results
         pred_result = None
-        if "virulencefinder" in pred_obj:
-            results = pred_obj["virulencefinder"]["results"]
-            species = list(results)
-            for assay, result in results[species[0]].items():
-                # skip non typing results
-                if not assay == "stx":
-                    continue
-
-                # if no stx gene was identified
-                if isinstance(result, str):
-                    continue
-
-                # take first result as the valid prediction
-                hit = next(iter(result.values()))
-                vir_gene = parse_vir_gene(hit)
-                gene = TypingResultGeneAllele(**vir_gene.model_dump())
-                pred_result = MethodIndex(
-                    type=TypingMethod.STX,
-                    software=Software.VIRFINDER,
-                    result=gene,
-                )
+        if "seq_regions" in pred_obj and "phenotypes" in pred_obj:
+            phenotypes = pred_obj.get("phenotypes", {})
+            seq_regions = pred_obj.get("seq_regions", {})
+    
+            stx_keys = [key for key in phenotypes if key.startswith("stx")]
+            if not stx_keys:
+                return None
+            
+            for stx_key in stx_keys:
+                pheno = phenotypes[stx_key]
+                function = pheno.get("function", "")
+                for region_key in pheno.get("seq_regions", []):
+                    region_info = seq_regions.get(region_key)
+                    if region_info:
+                        vir_gene = parse_vir_gene(region_info, function=function)
+                        # TODO cleanup data models. They are too inherited which makes it difficult to know what fields they contain
+                        gene = TypingResultGeneAllele(**vir_gene.model_dump())
+                        return MethodIndex(
+                            type=TypingMethod.STX,
+                            software=Software.VIRFINDER,
+                            result=gene,
+                        )
     return pred_result