+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 
 ### Changed
 
+ - Updated virulencefinder outputs for pytest
+ - Changed test fixtures for ecoli stx sample
+ - Removed `AmrFinderResistanceGene` as it confuses model
  - Set environment variables to avoid interactive prompts in `Dockerfile` by incl. TZ data
  - Updated `openjdk-17` to `openjdk-25` in `Dockerfile`